% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file 'filename'}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to read from.}
}
\value{
This function returns the data read from the specified file in table format.
}
\description{
This function reads in data from a comma separated value(csv) format file.
The output is returned in table format. If the named file does not exist, execution will stop and an error message displayed.
}
\examples{
fars_read('my_file.csv')

}
