% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{For a list / vector of years, create a filename for each year, read the corresponding data file,
add an additional column for the year and return the month and year columns from the dataset.
The function includes error trapping to capture invalid years and issue a warning.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of the years of interest.}
}
\value{
The function returns the month and year columns from the dataset for each specified year
}
\description{
For a list / vector of years, create a filename for each year, read the corresponding data file,
add an additional column for the year and return the month and year columns from the dataset.
The function includes error trapping to capture invalid years and issue a warning.
}
\examples{
fars_read_years(c(2013,2014,2015))

}
