% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Provide a monthly summary of the data for each year of interest.
Read data for each year of interest, bind rows into a single data frame,
group by year and month and summarise the results}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of the years of interest.}
}
\value{
A monthly summary of teh data for each year of interest.
}
\description{
Provide a monthly summary of the data for each year of interest.
Read data for each year of interest, bind rows into a single data frame,
group by year and month and summarise the results
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
