% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename for designated year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric year value}
}
\value{
this function returns a filename in a specific format
}
\description{
this function takes a numerical year as input and prints out a filename in a specified format.
a non-numeric year value will result ina n error.
}
\examples{
make_filename(2013)

}
