% Generated by roxygen2
\name{fars_read}
\alias{fars_read}

\title{Reaf file with FARS data}
\description{
This function reads data from .csv file, stored on disk, from the \strong{US
National Highway Traffic Safety Administration's} \emph{Fatality Analysis
Reporting System} (FARS), which is a nationwide census, providing the
American public yearly data, regarding fatal injuries suffered in motor
vehicle traffic crashes.
}
\usage{
fars_read(filename)
}

\arguments{
\item{filename}{A character string with the name of the file to read, see
notes.}
}

\details{
For more information, see:
\itemize{
  \item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
  \item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}
}
\value{
A data frame with data readed from the csv file, or an error if the
  file does not exists.
}

\author{
Williams Ojo
}
\note{
To generate file name use: \code{\link{make_filename}}
}

\seealso{
\link{make_filename}
}
\examples{
library(dplyr)
library(readr)
yr <- 2015
data <- yr \%>\%
  make_filename \%>\%
  fars_read
head(data)
}